requirements.txt
pandas>=2.0
numpy>=1.24
requests>=2.31
python-dotenv>=1.0
openpyxl>=3.1
streamlit>=1.30
plotly>=5.18

